+++ src/MatrixEquations.jl
@@ -8,6 +8,7 @@ using LinearAlgebra
 using LinearAlgebra: require_one_based_indexing
 import LinearAlgebra: mul!
 using LinearMaps
+#using MatrixPencils
 #using JLD